+++ qcfractal/storage_sockets/__init__.py
@@ -5,3 +5,5 @@ Importer for the DB socket class.
 __all__ = ["storage_socket_factory"]
 
 from .storage_socket import storage_socket_factory
+
+from .db_queries import TorsionDriveQueries